Follow PowerShell idioms for parameters
This change maintains the existing usage pattern (as described in the
README). It adds basic parameter validation, and also allows values to
be passed in as command-line parameters if desired.

The new usage patterns are documented in the README, and also documented
(with examples) using PowerShell's comment-based help syntax in the
script file.